Rename ToneMapper to PostProcesser
