+++ doc/classes/Viewport.xml
@@ -345,6 +345,9 @@
 		<member name="use_xr" type="bool" setter="set_use_xr" getter="is_using_xr" default="false">
 			If [code]true[/code], the viewport will use the primary XR interface to render XR output. When applicable this can result in a stereoscopic image and the resulting render being output to a headset.
 		</member>
+		<member name="viewport_mode" type="int" setter="set_viewport_mode" getter="get_viewport_mode" enum="Viewport.ViewportMode" default="0">
+			Change global mode of viewport and output type.
+		</member>
 		<member name="vrs_mode" type="int" setter="set_vrs_mode" getter="get_vrs_mode" enum="Viewport.VRSMode" default="0">
 			The Variable Rate Shading (VRS) mode that is used for this viewport. Note, if hardware does not support VRS this property is ignored.
 		</member>
@@ -441,6 +444,15 @@
 		</constant>
 		<constant name="SCREEN_SPACE_AA_MAX" value="2" enum="ScreenSpaceAA">
 			Represents the size of the [enum ScreenSpaceAA] enum.
+		</constant>
+		<constant name="2D_AND_3D" value="0" enum="ViewportMode">
+			Standard render pipeline, serve a texture ready for screen
+		</constant>
+		<constant name="3D" value="1" enum="ViewportMode">
+			Render with no linearization nor tonemapping
+		</constant>
+		<constant name="XR" value="2" enum="ViewportMode">
+			Render XR output
 		</constant>
 		<constant name="RENDER_INFO_OBJECTS_IN_FRAME" value="0" enum="RenderInfo">
 			Amount of objects in frame.
